+++ CHANGES.md
@@ -11,6 +11,11 @@
 
 ## develop
 
+- [CHANGE] `@available(*, unavailable)` は廃止になるため削除する
+    - Swift 5.9 以降 `@available(*, unavailable)` が禁止された
+    - Sora iOS SDK では廃止となったプロパティに対して `@available(*, unavailable)` を付与していたが、削除した
+    - @torikizi
+
 ## 2023.2.0
 
 - [UPDATE] システム条件を変更する